twenty four migration

+++ StoredProcedure/SpReportByAll.cs
@@ -6,7 +6,7 @@
         public double inward { get; set; }
         public double Outward { get; set; }
         public double closingbalance { get; set; }
-        public Int64 ItemCode { get; set; }
+        public Int32 ItemCode { get; set; }
         public string ItemName { get; set; }
        public double Freeze { get; set; }
         public double ReturnedFreshStock { get; set; }
